+++ frontend/app/src/store/defi/types.ts
@@ -49,8 +49,10 @@ export interface DefiState {
   allProtocols: AllDefiProtocols;
   compoundBalances: CompoundBalances;
   compoundHistory: CompoundHistory;
-  yearnVaultsHistory: YearnVaultsHistory;
   yearnVaultsBalances: YearnVaultsBalances;
+  yearnVaultsHistory: YearnVaultsHistory;
+  yearnVaultsV2Balances: YearnVaultsBalances;
+  yearnVaultsV2History: YearnVaultsHistory;
   uniswapBalances: UniswapBalances;
   uniswapTrades: DexTrades;
   uniswapEvents: UniswapEvents;
@@ -213,6 +215,7 @@ interface LendingHistoryExtras {
   readonly makerdao: MakerDAOLendingHistoryExtras;
   readonly compound: HistoryExtras<CompoundEventType>;
   readonly yearn_vaults: HistoryExtras<YearnEventType>;
+  readonly yearn_vaults_v2: HistoryExtras<YearnEventType>;
   readonly uniswap: {};
 }